--- a/go.mod
+++ b/go.mod
@@ -18,7 +18,7 @@ go 1.18
 
 require (
 	github.com/anuraaga/go-modsecurity v0.0.0-20220824035035-b9a4099778df
-	github.com/corazawaf/libinjection-go v0.0.0-20220207031228-44e9c4250eb5
+	github.com/corazawaf/libinjection-go v0.0.0-20220909190158-227e7e772cef
 	github.com/foxcpp/go-mockdns v1.0.0
 	github.com/magefile/mage v1.13.0
 	github.com/petar-dambovaliev/aho-corasick v0.0.0-20211021192214-5ab2d9280aa9