--- a/testing/coreruleset_test.go
+++ b/testing/coreruleset_test.go
@@ -65,6 +65,7 @@ func BenchmarkCRSCompilation(b *testing.B) {
 func BenchmarkCRSSimpleGET(b *testing.B) {
 	waf := crsWAF(b)
 
+	b.ResetTimer() // only benchmark execution, not compilation
 	for i := 0; i < b.N; i++ {
 		tx := waf.NewTransaction(context.Background())
 		tx.ProcessConnection("127.0.0.1", 8080, "127.0.0.1", 8080)