Let's get this out

After
```
Preparing CRS...
CRS PATH: /var/folders/vx/38td0k79201f9zn_myzbqx_40000gn/T/crs3387301874
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/testing
BenchmarkCRSCompilation
BenchmarkCRSCompilation-10    	      19	  60148535 ns/op
BenchmarkCRSSimpleGET
BenchmarkCRSSimpleGET-10      	    1219	    967255 ns/op
BenchmarkCRSSimplePOST
BenchmarkCRSSimplePOST-10     	     735	   1610899 ns/op
PASS
```

Before
```
Preparing CRS...
CRS PATH: /var/folders/vx/38td0k79201f9zn_myzbqx_40000gn/T/crs3711353421
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/testing
BenchmarkCRSCompilation
BenchmarkCRSCompilation-10    	      18	  59931525 ns/op
BenchmarkCRSSimpleGET
BenchmarkCRSSimpleGET-10      	     966	   1306669 ns/op
BenchmarkCRSSimplePOST
BenchmarkCRSSimplePOST-10     	     530	   2177657 ns/op
PASS
```